+++ src/main/java/org/csanchez/jenkins/plugins/kubernetes/ContainerTemplate.java
@@ -8,6 +8,8 @@ import java.util.List;
 import org.apache.commons.lang.StringUtils;
 import org.csanchez.jenkins.plugins.kubernetes.model.TemplateEnvVar;
 import org.jenkinsci.Symbol;
+import org.kohsuke.accmod.Restricted;
+import org.kohsuke.accmod.restrictions.DoNotUse;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
 
@@ -16,6 +18,8 @@ import com.google.common.base.Preconditions;
 import hudson.Extension;
 import hudson.model.AbstractDescribableImpl;
 import hudson.model.Descriptor;
+import hudson.model.DescriptorVisibilityFilter;
+import jenkins.model.Jenkins;
 
 public class ContainerTemplate extends AbstractDescribableImpl<ContainerTemplate> implements Serializable {
 
@@ -218,5 +222,32 @@ public class ContainerTemplate extends AbstractDescribableImpl<ContainerTemplate
         public String getDisplayName() {
             return "Container Template";
         }
+
+        @SuppressWarnings("unused") // Used by jelly
+        @Restricted(DoNotUse.class) // Used by jelly
+        public List<? extends Descriptor> getEnvVarsDescriptors() {
+            return DescriptorVisibilityFilter.apply(null, Jenkins.getInstance().getDescriptorList(TemplateEnvVar.class));
+        }
+    }
+
+    @Override
+    public String toString() {
+        return "ContainerTemplate{" +
+                (name == null ? "" : "name='" + name + '\'') +
+                (image == null ? "" : ", image='" + image + '\'') +
+                (!privileged ? "" : ", privileged=" + privileged) +
+                (!alwaysPullImage ? "" : ", alwaysPullImage=" + alwaysPullImage) +
+                (workingDir == null ? "" : ", workingDir='" + workingDir + '\'') +
+                (command == null ? "" : ", command='" + command + '\'') +
+                (args == null ? "" : ", args='" + args + '\'') +
+                (!ttyEnabled ? "" : ", ttyEnabled=" + ttyEnabled) +
+                (resourceRequestCpu == null ? "" : ", resourceRequestCpu='" + resourceRequestCpu + '\'') +
+                (resourceRequestMemory == null ? "" : ", resourceRequestMemory='" + resourceRequestMemory + '\'') +
+                (resourceLimitCpu == null ? "" : ", resourceLimitCpu='" + resourceLimitCpu + '\'') +
+                (resourceLimitMemory == null ? "" : ", resourceLimitMemory='" + resourceLimitMemory + '\'') +
+                (envVars == null || envVars.isEmpty() ? "" : ", envVars=" + envVars) +
+                (ports == null || ports.isEmpty() ? "" : ", ports=" + ports) +
+                (livenessProbe == null ? "" : ", livenessProbe=" + livenessProbe) +
+                '}';
     }
 }