+++ src/main/java/org/csanchez/jenkins/plugins/kubernetes/KubernetesCloud.java
@@ -25,6 +25,8 @@ import javax.servlet.ServletException;
 
 import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.lang.StringUtils;
+import org.jenkinsci.plugins.plaincredentials.StringCredentials;
+import org.jenkinsci.plugins.plaincredentials.impl.StringCredentialsImpl;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
 import org.kohsuke.stapler.QueryParameter;
@@ -42,8 +44,12 @@ import com.google.common.collect.ImmutableMap;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import org.csanchez.jenkins.plugins.kubernetes.pipeline.PodTemplateMap;
+
 import hudson.Extension;
 import hudson.Util;
+import hudson.init.InitMilestone;
+import hudson.init.Initializer;
 import hudson.model.Computer;
 import hudson.model.Descriptor;
 import hudson.model.Label;
@@ -64,7 +70,7 @@ import jenkins.model.JenkinsLocationConfiguration;
 /**
  * Kubernetes cloud provider.
  *
- * Starts slaves in a Kubernetes cluster using defined Docker templates for each label.
+ * Starts agents in a Kubernetes cluster using defined Docker templates for each label.
  *
  * @author Carlos Sanchez carlos@apache.org
  */
@@ -77,6 +83,7 @@ public class KubernetesCloud extends Cloud {
 
     public static final String JNLP_NAME = "jnlp";
     /** label for all pods started by the plugin */
+    @Deprecated
     public static final Map<String, String> DEFAULT_POD_LABELS = ImmutableMap.of("jenkins", "slave");
 
     /** Default timeout for idle workers that don't correctly indicate exit. */
@@ -84,7 +91,8 @@ public class KubernetesCloud extends Cloud {
 
     private String defaultsProviderTemplate;
 
-    private List<PodTemplate> templates = new ArrayList<PodTemplate>();
+    @Nonnull
+    private List<PodTemplate> templates = new ArrayList<>();
     private String serverUrl;
     @CheckForNull
     private String serverCertificate;
@@ -101,6 +109,7 @@ public class KubernetesCloud extends Cloud {
     private int retentionTimeout = DEFAULT_RETENTION_TIMEOUT_MINUTES;
     private int connectTimeout;
     private int readTimeout;
+    private Map<String, String> labels;
 
     private transient KubernetesClient client;
     private int maxRequestsPerHost;
@@ -168,13 +177,25 @@ public class KubernetesCloud extends Cloud {
         this.defaultsProviderTemplate = defaultsProviderTemplate;
     }
 
+    @Nonnull
     public List<PodTemplate> getTemplates() {
         return templates;
     }
 
+    /**
+     * Returns all pod templates for this cloud including the dynamic ones.
+     * @return all pod templates for this cloud including the dynamic ones.
+     */
+    @Nonnull
+    public List<PodTemplate> getAllTemplates() {
+        List<PodTemplate> podTemplates = new ArrayList<>(PodTemplateMap.get().getTemplates(this));
+        podTemplates.addAll(templates);
+        return Collections.unmodifiableList(podTemplates);
+    }
+
     @DataBoundSetter
     public void setTemplates(@Nonnull List<PodTemplate> templates) {
-        this.templates = templates;
+        this.templates = new ArrayList<>(templates);
     }
 
     public String getServerUrl() {
@@ -305,6 +326,17 @@ public class KubernetesCloud extends Cloud {
         return connectTimeout;
     }
 
+    /**
+     * Labels for all pods started by the plugin
+     */
+    public Map<String, String> getLabels() {
+        return labels == null ? Collections.emptyMap() : labels;
+    }
+
+    public void setLabels(Map<String, String> labels) {
+        this.labels = labels;
+    }
+
     @DataBoundSetter
     public void setMaxRequestsPerHostStr(String maxRequestsPerHostStr) {
         try  {
@@ -348,7 +380,7 @@ public class KubernetesCloud extends Cloud {
 
     Map<String, String> getLabelsMap(Set<LabelAtom> labelSet) {
         ImmutableMap.Builder<String, String> builder = ImmutableMap.<String, String> builder();
-        builder.putAll(DEFAULT_POD_LABELS);
+        builder.putAll(getLabels());
         if (!labelSet.isEmpty()) {
             for (LabelAtom label: labelSet) {
                 builder.put(getIdForLabel(label), "true");
@@ -417,7 +449,7 @@ public class KubernetesCloud extends Cloud {
             templateNamespace = client.getNamespace();
         }
 
-        PodList slaveList = client.pods().inNamespace(templateNamespace).withLabels(DEFAULT_POD_LABELS).list();
+        PodList slaveList = client.pods().inNamespace(templateNamespace).withLabels(getLabels()).list();
         List<Pod> slaveListItems = slaveList.getItems();
 
         Map<String, String> labelsMap = getLabelsMap(template.getLabelSet());
@@ -426,16 +458,16 @@ public class KubernetesCloud extends Cloud {
 
         if (slaveListItems != null && containerCap <= slaveListItems.size()) {
             LOGGER.log(Level.INFO,
-                    "Total container cap of {0} reached, not provisioning: {1} running or errored in namespace {2}",
-                    new Object[] { containerCap, slaveListItems.size(), client.getNamespace() });
+                    "Total container cap of {0} reached, not provisioning: {1} running or errored in namespace {2} with Kubernetes labels {3}",
+                    new Object[] { containerCap, slaveListItems.size(), client.getNamespace(), getLabels() });
             return false;
         }
 
         if (namedListItems != null && slaveListItems != null && template.getInstanceCap() <= namedListItems.size()) {
             LOGGER.log(Level.INFO,
-                    "Template instance cap of {0} reached for template {1}, not provisioning: {2} running or errored in namespace {3} with label {4}",
+                    "Template instance cap of {0} reached for template {1}, not provisioning: {2} running or errored in namespace {3} with label \"{4}\" and Kubernetes labels {5}",
                     new Object[] { template.getInstanceCap(), template.getName(), slaveListItems.size(),
-                            client.getNamespace(), label == null ? "" : label.toString() });
+                            client.getNamespace(), label == null ? "" : label.toString(), labelsMap });
             return false; // maxed out
         }
         return true;
@@ -452,7 +484,7 @@ public class KubernetesCloud extends Cloud {
      * @return the template
      */
     public PodTemplate getTemplate(@CheckForNull Label label) {
-        return PodTemplateUtils.getTemplateByLabel(label, templates);
+        return PodTemplateUtils.getTemplateByLabel(label, getAllTemplates());
     }
 
     /**
@@ -462,7 +494,8 @@ public class KubernetesCloud extends Cloud {
      */
     public ArrayList<PodTemplate> getMatchingTemplates(@CheckForNull Label label) {
         ArrayList<PodTemplate> podList = new ArrayList<PodTemplate>();
-        for (PodTemplate t : templates) {
+        List<PodTemplate> podTemplates = getAllTemplates();
+        for (PodTemplate t : podTemplates) {
             if ((label == null && t.getNodeUsageMode() == Node.Mode.NORMAL) || (label != null && label.matches(t.getLabelSet()))) {
                 podList.add(t);
             }
@@ -488,11 +521,39 @@ public class KubernetesCloud extends Cloud {
         this.templates.remove(t);
     }
 
+    /**
+     * Add a dynamic pod template. Won't be displayed in UI, and persisted separately from the cloud instance.
+     * @param t the template to add
+     */
+    public void addDynamicTemplate(PodTemplate t) {
+        PodTemplateMap.get().addTemplate(this, t);
+    }
+
+    /**
+     * Remove a dynamic pod template.
+     * @param t the template to remove
+     */
+    public void removeDynamicTemplate(PodTemplate t) {
+        PodTemplateMap.get().removeTemplate(this, t);
+    }
+
     @Extension
     public static class DescriptorImpl extends Descriptor<Cloud> {
         @Override
         public String getDisplayName() {
             return "Kubernetes";
+        }
+
+        @Initializer(before = InitMilestone.PLUGINS_STARTED)
+        public static void addAliases() {
+            Jenkins.XSTREAM2.addCompatibilityAlias(
+                    "org.csanchez.jenkins.plugins.kubernetes.OpenShiftBearerTokenCredentialImpl",
+                    org.jenkinsci.plugins.kubernetes.credentials.OpenShiftBearerTokenCredentialImpl.class);
+            Jenkins.XSTREAM2.addCompatibilityAlias(
+                    "org.csanchez.jenkins.plugins.kubernetes.OpenShiftTokenCredentialImpl",
+                    StringCredentialsImpl.class);
+            Jenkins.XSTREAM2.addCompatibilityAlias("org.csanchez.jenkins.plugins.kubernetes.ServiceAccountCredential",
+                    org.jenkinsci.plugins.kubernetes.credentials.FileSystemServiceAccountCredential.class);
         }
 
         public FormValidation doTestConnection(@QueryParameter String name, @QueryParameter String serverUrl, @QueryParameter String credentialsId,
@@ -525,19 +586,20 @@ public class KubernetesCloud extends Cloud {
         }
 
         public ListBoxModel doFillCredentialsIdItems(@QueryParameter String serverUrl) {
-            return new StandardListBoxModel()
-                    .withEmptySelection()
-                    .withMatching(
+            return new StandardListBoxModel().withEmptySelection() //
+                    .withMatching( //
                             CredentialsMatchers.anyOf(
                                     CredentialsMatchers.instanceOf(StandardUsernamePasswordCredentials.class),
                                     CredentialsMatchers.instanceOf(TokenProducer.class),
-                                    CredentialsMatchers.instanceOf(StandardCertificateCredentials.class)
-                            ),
-                            CredentialsProvider.lookupCredentials(StandardCredentials.class,
-                                    Jenkins.getInstance(),
-                                    ACL.SYSTEM,
+                                    CredentialsMatchers.instanceOf(
+                                            org.jenkinsci.plugins.kubernetes.credentials.TokenProducer.class),
+                                    CredentialsMatchers.instanceOf(StandardCertificateCredentials.class),
+                                    CredentialsMatchers.instanceOf(StringCredentials.class)), //
+                            CredentialsProvider.lookupCredentials(StandardCredentials.class, //
+                                    Jenkins.getInstance(), //
+                                    ACL.SYSTEM, //
                                     serverUrl != null ? URIRequirementBuilder.fromUri(serverUrl).build()
-                                                      : Collections.EMPTY_LIST
+                                            : Collections.EMPTY_LIST //
                             ));
 
         }
@@ -566,6 +628,9 @@ public class KubernetesCloud extends Cloud {
 
         if (maxRequestsPerHost == 0) {
             maxRequestsPerHost = DEFAULT_MAX_REQUESTS_PER_HOST;
+        }
+        if (labels == null) {
+            labels = DEFAULT_POD_LABELS;
         }
         return this;
     }