+++ src/main/java/org/csanchez/jenkins/plugins/kubernetes/KubernetesComputer.java
@@ -27,7 +27,7 @@ public class KubernetesComputer extends AbstractCloudComputer<KubernetesSlave> {
     public void taskCompleted(Executor executor, Queue.Task task, long durationMS) {
         LOGGER.log(Level.FINE, " Computer " + this + " taskCompleted");
 
-        // May take the slave offline and remove it, in which case getNode()
+        // May take the agent offline and remove it, in which case getNode()
         // above would return null and we'd not find our DockerSlave anymore.
         super.taskCompleted(executor, task, durationMS);
     }