+++ src/test/java/org/csanchez/jenkins/plugins/kubernetes/KubernetesTestUtil.java
@@ -31,6 +31,7 @@ import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
 import java.security.UnrecoverableKeyException;
 import java.security.cert.CertificateEncodingException;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ForkJoinPool;
@@ -41,6 +42,9 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import org.apache.commons.lang.StringUtils;
+
+import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.Maps;
 
 import io.fabric8.kubernetes.api.model.NamespaceBuilder;
 import io.fabric8.kubernetes.api.model.Pod;
@@ -61,9 +65,17 @@ public class KubernetesTestUtil {
     private static final String DEFAULT_TESTING_NAMESPACE = "kubernetes-plugin-test";
     public static String testingNamespace;
 
-    public static KubernetesCloud setupCloud() throws UnrecoverableKeyException, CertificateEncodingException,
-            NoSuchAlgorithmException, KeyStoreException, IOException {
+    private static final String BRANCH_NAME = System.getenv("BRANCH_NAME");
+    private static final String BUILD_NUMBER = System.getenv("BUILD_NUMBER");
+    private static Map<String, String> DEFAULT_LABELS = ImmutableMap.of("BRANCH_NAME",
+            BRANCH_NAME == null ? "undefined" : BRANCH_NAME, "BUILD_NUMBER",
+            BUILD_NUMBER == null ? "undefined" : BUILD_NUMBER);
+
+    public static KubernetesCloud setupCloud(Object test) throws UnrecoverableKeyException,
+            CertificateEncodingException, NoSuchAlgorithmException, KeyStoreException, IOException {
         KubernetesCloud cloud = new KubernetesCloud("kubernetes");
+        // unique labels per test
+        cloud.setLabels(getLabels(test));
         KubernetesClient client = cloud.connect();
 
         // Run in our own testing namespace
@@ -103,6 +115,12 @@ public class KubernetesTestUtil {
         } catch (Exception e) {
             assumeNoException(e);
         }
+    }
+
+    public static Map<String, String> getLabels(Object o) {
+        HashMap<String, String> l = Maps.newHashMap(DEFAULT_LABELS);
+        l.put("class", o.getClass().getSimpleName());
+        return l;
     }
 
     /**